+++ packages/stg-definitions-v2/src/constant.ts
@@ -238,7 +238,17 @@ export const ASSETS: Record<TokenName, AssetConfig> = {
             [EndpointId.ARBSEP_V2_TESTNET]: {
                 type: StargateType.Native,
             },
+            [EndpointId.BL3_V2_TESTNET]: {
+                symbol: 'WETH',
+                name: 'WETH',
+                type: StargateType.Oft,
+            },
             [EndpointId.KLAYTN_V2_TESTNET]: {
+                symbol: 'WETH',
+                name: 'WETH',
+                type: StargateType.Oft,
+            },
+            [EndpointId.ODYSSEY_V2_TESTNET]: {
                 symbol: 'WETH',
                 name: 'WETH',
                 type: StargateType.Oft,
@@ -363,12 +373,20 @@ export const ASSETS: Record<TokenName, AssetConfig> = {
             [EndpointId.ARBSEP_V2_TESTNET]: {
                 type: StargateType.Pool,
             },
+            [EndpointId.BL3_V2_TESTNET]: {
+                type: StargateType.Oft,
+                address: '0x164A2dE1bc5dc56F329909F7c97Bae929CaE557B',
+            },
             [EndpointId.BSC_V2_TESTNET]: {
                 address: '0xe37Bdc6F09DAB6ce6E4eBC4d2E72792994Ef3765',
                 type: StargateType.Pool,
             },
             [EndpointId.KLAYTN_V2_TESTNET]: {
                 type: StargateType.Oft,
+            },
+            [EndpointId.ODYSSEY_V2_TESTNET]: {
+                type: StargateType.Oft,
+                address: '0x880f614c729c622c727f3625b263a36c0b89167F',
             },
             [EndpointId.OPTSEP_V2_TESTNET]: {
                 type: StargateType.Pool,
@@ -510,9 +528,17 @@ export const ASSETS: Record<TokenName, AssetConfig> = {
                 address: '0x3253a335E7bFfB4790Aa4C25C4250d206E9b9773',
                 type: StargateType.Pool,
             },
+            [EndpointId.BL3_V2_TESTNET]: {
+                type: StargateType.Oft,
+                address: '0x015fd589F4f1A33ce4487E12714e1B15129c9329',
+            },
             [EndpointId.KLAYTN_V2_TESTNET]: {
                 symbol: 'USDC.e',
                 type: StargateType.Oft,
+            },
+            [EndpointId.ODYSSEY_V2_TESTNET]: {
+                type: StargateType.Oft,
+                address: '0x88853D410299BCBfE5fCC9Eef93c03115E908279',
             },
             [EndpointId.OPTSEP_V2_TESTNET]: {
                 address: '0x488327236B65C61A6c083e8d811a4E0D3d1D4268',
@@ -631,8 +657,10 @@ export const OFT_WRAPPER: OftWrapperConfig = {
         // TESTNET
         //
         [EndpointId.ARBSEP_V2_TESTNET]: {},
+        [EndpointId.BL3_V2_TESTNET]: {},
         [EndpointId.BSC_V2_TESTNET]: {},
         [EndpointId.KLAYTN_V2_TESTNET]: {},
+        [EndpointId.ODYSSEY_V2_TESTNET]: {},
         [EndpointId.OPTSEP_V2_TESTNET]: {},
         [EndpointId.SEPOLIA_V2_TESTNET]: {},
 
@@ -1495,11 +1523,19 @@ export const NETWORKS: NetworksConfig = {
         creditMessaging: DEFAULT_CREDIT_MESSAGING_NETWORK_CONFIG,
         tokenMessaging: DEFAULT_TOKEN_MESSAGING_NETWORK_CONFIG,
     },
+    [EndpointId.BL3_V2_TESTNET]: {
+        creditMessaging: DEFAULT_CREDIT_MESSAGING_NETWORK_CONFIG,
+        tokenMessaging: DEFAULT_TOKEN_MESSAGING_NETWORK_CONFIG,
+    },
     [EndpointId.BSC_V2_TESTNET]: {
         creditMessaging: DEFAULT_CREDIT_MESSAGING_NETWORK_CONFIG,
         tokenMessaging: DEFAULT_TOKEN_MESSAGING_NETWORK_CONFIG,
     },
     [EndpointId.KLAYTN_V2_TESTNET]: {
+        creditMessaging: DEFAULT_CREDIT_MESSAGING_NETWORK_CONFIG,
+        tokenMessaging: DEFAULT_TOKEN_MESSAGING_NETWORK_CONFIG,
+    },
+    [EndpointId.ODYSSEY_V2_TESTNET]: {
         creditMessaging: DEFAULT_CREDIT_MESSAGING_NETWORK_CONFIG,
         tokenMessaging: DEFAULT_TOKEN_MESSAGING_NETWORK_CONFIG,
     },