+++ packages/stg-evm-v2/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @stargatefinance/stg-evm-v2
 
+## 1.2.3
+
+### Patch Changes
+
+- 974c8e6: Story Odyssey Testnet
+- Updated dependencies [974c8e6]
+  - @stargatefinance/stg-definitions-v2@1.3.3
+
+## 1.2.2
+
+### Patch Changes
+
+- 253a1a8: bl3 testnet
+- da1a8b9: degen mainnet
+- Updated dependencies [253a1a8]
+- Updated dependencies [da1a8b9]
+  - @stargatefinance/stg-definitions-v2@1.3.2
+
 ## 1.2.1
 
 ### Patch Changes